server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pillow==10.1.0
torch==2.1.0
diffusers==0.23.1
transformers==4.34.1
accelerate==0.23.0
safetensors==0.4.0
xformers==0.0.22.post7
python-multipart==0.0.6
